pyproject.toml
[project]
name = "iot-app"
version = "0.1.0"
description = "Dummy Modbus TCP temperature sensor for IoT testing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pymodbus>=3.11.4",
]

[project.scripts]
iot-dummy = "main:main"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "ruff>=0.14.13",
]

[tool.setuptools]
package-dir = { "" = "src" }
py-modules = [
    "main",
    "modbus_server",
    "mqtt_server",
]
